added error handling responses for employee rest controller

+++ employee-skills-tracking-application/employee-skills-tracking-application/src/main/java/com/example/employeeskillstrackingapplication/EmployeeService.java
@@ -2,36 +2,61 @@ package com.example.employeeskillstrackingapplication;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import exceptions.EmployeeNotFoundException;
+import exceptions.InvalidIdException;
+
 @Service
 public class EmployeeService {
-	
+
 	@Autowired
 	EmployeeRepository employeeRepository;
-	
-	public List<Employee> getAllEmployees(){
+
+	public List<Employee> getAllEmployees() {
 		List<Employee> allEmployees = new ArrayList<>();
 		employeeRepository.findAll().forEach(employee -> allEmployees.add(employee));
 		return allEmployees;
 	}
-	
-	public Employee getEmployeeById(int id) {
-		return employeeRepository.findById(id).get();
+
+	public Employee getEmployeeById(Integer id) {
+		Optional<Employee> employeeOp = employeeRepository.findById(id);
+		Employee requestedEmployee = employeeOp
+				.orElseThrow(() -> new EmployeeNotFoundException("Perficient Employee Not Found"));
+		return requestedEmployee;
+	}
+
+	public Employee save(Employee employee) {
+		if (isValidEmployee(employee)) {
+			return employeeRepository.save(employee);
+		}
+		return null;
 	}
 	
-	public Employee saveOrUpdate(Employee employee) {
-		return employeeRepository.save(employee);
+	public Employee update(Employee employee) {
+		int employeeId = employee.getEmployeeId();
+		if (getEmployeeById(employeeId) == null) {
+			throw new EmployeeNotFoundException("Perficient Employee Not Found");
+		}
+		return save(employee);
 	}
-	
+
 	public void delete(int id) {
+		if (getEmployeeById(id) == null) {
+			throw new EmployeeNotFoundException("Perficient Employee Not Found");
+		}
 		employeeRepository.deleteById(id);
 	}
-	
-	
-	
-	
+
+	private boolean isValidEmployee(Employee employee) {
+		if (employee.getFirstName() != null && employee.getLastName() != null && employee.getPosition() != null) {
+			return true;
+		}
+
+		return false;
+	}
 
 }
